Catch development store errors during init (#1281)
* Remove unused imports

* Use more descriptive describe blocks

We're about to add another case to this so I'm cleaning it up ahead of
time to make the functional diff easier to read.

* Catch access denied error during init flow

Developers are hitting this error when they try to create a new
Hydrogen storefront on a Development Store. This catches that error and
provides some better feedback on why it failed.